+++ modules/alloy/src/main/ts/ephox/alloy/behaviour/dragging/DraggingBranches.ts
@@ -1,10 +1,13 @@
-import MouseDragging from '../../dragging/mouse/MouseDragging';
-import TouchDragging from '../../dragging/touch/TouchDragging';
+import * as MouseDragging from '../../dragging/mouse/MouseDragging';
+import * as MouseOrTouchDragging from '../../dragging/mouseortouch/MouseOrTouchDragging';
+import * as TouchDragging from '../../dragging/touch/TouchDragging';
 
-const mouse = MouseDragging;
-const touch = TouchDragging;
+const mouse = MouseDragging.schema;
+const touch = TouchDragging.schema;
+const mouseOrTouch = MouseOrTouchDragging.schema;
 
 export {
   mouse,
-  touch
+  touch,
+  mouseOrTouch
 };