+++ modules/alloy/changelog.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 # [5.1.0] - 2019-12-17
 
+### Added
+- Added new `mouseOrTouch` mode to dragging to support dragging for both mouse and touch events.
+
 ### Changed
 - Changed touch/mouse event handling to work with hybrid devices that accept both mouse and touch events.
 - Changed `AlloyEvents.runActionExtra()` to pass the simulated event to the callback.